requirements.txt
asgiref
Django
django-ckeditor
django-cleanup
django-filter
djangorestframework
django-jazzmin
django-js-asset
django-phonenumber-field
django-resized
phonenumbers
Pillow
pytz
xmltodict